Update PII_stata_scan.do
Updating the PII scanner to reflect small bug fixes and larger updates to reduce the number of false positives. Specifically, allows the user to split the search strings between search strings that are likely to indicate direct PII and those that may are less likely and more common in variable names. It applies more stringent rules to the latter to reduce total output.